third_party/WebKit/LayoutTests/virtual/service-worker-navigation-preload/http/tests/serviceworker/chromium/navigation-preload-origin-trial-methods-expected.txt

CONSOLE WARNING: line 13: This test relies on console message comparison so there can be different -expected.txt files for virtual test suites.
CONSOLE MESSAGE: line 68: --normal SW--
disable() resolved with: undefined
enable() resolved with: undefined
getState() resolved with: {"enabled":true,"headerValue":"world"}
setHeaderValue() resolved with: undefined
log from SW
Activate event handler
disable() resolved with: undefined
enable() resolved with: undefined
getState() resolved with: {"enabled":true,"headerValue":"true"}
setHeaderValue() resolved with: undefined
Fetch event handler
event.preloadResponse resolved with: [object Response]
Text of preloadResponse: [Service-Worker-Navigation-Preload header set: hello]

CONSOLE MESSAGE: line 79: --Origin-Trial enabled SW--
disable() resolved with: undefined
enable() resolved with: undefined
getState() resolved with: {"enabled":true,"headerValue":"world"}
setHeaderValue() resolved with: undefined
log from SW
Activate event handler
disable() resolved with: undefined
enable() resolved with: undefined
getState() resolved with: {"enabled":true,"headerValue":"true"}
setHeaderValue() resolved with: undefined
Fetch event handler
event.preloadResponse resolved with: [object Response]
Text of preloadResponse: [Service-Worker-Navigation-Preload header set: hello]

This is a testharness.js-based test.
PASS Calling Navigation preload related methods for normal SW. 
PASS Calling Navigation preload related methods for Origin-Trial enabled SW. 
Harness: the test ran to completion.
